requirements.txt
# Python dependencies
paho-mqtt==1.6.1
boto3==1.34.24
meshtastic>=2.2.0
